requirements.txt
asgiref==3.3.1
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
Django==3.0.4
django-cors-headers==3.1.0
djangorestframework==3.14.0
filelock==3.8.0
fonttools==4.38.0
gunicorn==20.0.4
h11==0.13.0
idna==2.10
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.4
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.2
pycparser==2.21
pyfcm==1.5.4
pyparsing==3.0.9
pypng==0.20220715.0
PyQRCode==1.2.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2020.5
pyzbar==0.1.9
razorpay==1.2.0
requests==2.25.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.2
validators==0.20.0
virtualenv==20.16.5
wincertstore==0.2
wsproto==1.2.0
